+++ packages/discord-utilities/src/index.ts
@@ -1,4 +1,3 @@
-export type { RequiredIf } from './lib/_utils';
 export * from './lib/InteractionOptionResolver';
 export * from './lib/limits';
 export * from './lib/regexes';
